API_COCKROACHDB/requirements.txt
amqp==2.3.2
aniso8601==3.0.0
arrow==0.12.1
astroid==2.0.4
atomicwrites==1.2.1
attrs==17.4.0
Babel==2.6.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
click==6.7
coverage==4.5.1
Cython==0.28.1
Flask==0.12.2
Flask-Celery==2.4.3
Flask-Cors==3.0.3
Flask-InfluxDB==0.1
Flask-JWT-Extended==3.7.2
Flask-Redis==0.3.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SocketIO==3.0.2
flask-swagger-ui==3.6.0
Flask-WTF==0.14.2
flower==0.9.2
gevent==1.3.6
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==19.9.0
idna==2.7
influxdb==5.2.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
lazy-object-proxy==1.3.1
ldap3==2.4.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.1.0
passlib==1.7.1
pluggy==0.6.0
psycopg2==2.7.5
py==1.5.3
pyasn1==0.4.2
PyJWT==1.4.2
pylint==2.1.1
pymongo==3.6.1
pytest==3.5.0
pytest-cov==2.5.1
pytest-flask==0.10.0
python-dateutil==2.7.2
python-dotenv==0.8.2
python-engineio==2.2.0
python-memcached==1.59
python-socketio==2.0.0
pytz==2018.3
PyYAML==3.13
redis==2.10.6
requests==2.19.1
six==1.11.0
socketIO-client==0.7.2
tornado==5.1.1
typed-ast==1.1.0
uniqid==0.0.1
urllib3==1.23
vine==1.1.4
websocket-client==0.54.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
